Create cleanup-old-user-profile.ps1
On a Windows machine, this script can be used to entirely delete a local user profile. Before starting the removal procedure, the script first asks the user for approval.

The script gets the user account and deletes it from the local account database as soon as the user confirms. The user's profile directory is then attempted to be deleted, and as it does so, a progress bar showing the removal's status and progress is displayed. The script advances to the following stage if the user profile directory cannot be located.

Once the user's profile has been deleted from the registry, the script shows a message letting the user know whether the profile was located and deleted. The script then displays a notice stating that the user profile has been entirely deleted.

For the script to work, the user must identify the local user account whose profile has to be deleted. For instance, the user would use Remove-LocalUserProfileCompletely -Username "JohnDoe" to delete the local user profile for the user "JohnDoe".

Karanvir is the author of the script.